NEW Allow silverstripe/errorpage to static publish
When using the error page module to create custom content error pages
within the CMS (for e.g. HTTP 404 & 500) the default is to also static
publish these files so the web server can reference them directly. This
requires an unauthenticated execution of the controller in a mocked
request in order to get all the page HTML without revealing any user
based customised data that may be present. SAMLMiddleware will currently
redirect this request to the IdP for authentication when enabled,
essentially preventing error pages from being published, ever (in the
default configuration).

Hooks to customise the environment before dispatching the mocked request
to the error page controller have recently been added to the
ErrorPage, allowing us to disable the middleware & re-set it after the
mock request completes. We can apply this via config only if the module
exists.