Remove redundant code
SilverStripe 2 & 3 used an abstract class to as an interface to
implement authenticators. Usable implementors could identify themselves
via get_name and a designate a log in Form class via get_login_form

Silverstripe CMS 4 completely refactored the authentication API and
introduced an Authenticator interface defining a method getLoginHandler.
There is also a LoginForm abstract class that implements a method
getAuthenticatorName for the associated Authenticator.

These old static methods are no longer requird, and the private property
is not referred to anywhere within the class. I suspect it might be a
typographical omission where it was supposed to be private static $name,
a "default" for the config referred to in get_name (also removed).