Add tests
Most of the code in the module is untested. This has led to odd bugs,
such as SAMLController::getRedirect using the $this->owner-> reference
which will cause a fatal error because it is not an instance of
`Extension`.

Writing tests has interesting side effects such as ensuring that the
unique ID used in logs from SAMLController::acs is applied consistently,
which aids debugging should it need to happen. Logging is now also
applied more consistently, rather than simply assigning a message to the
login-form presented to the user.

SAMLController::getForm was changed to private as being public was
misleading for its purpose. The SAMLController never renders output (it
always redirects), and is not responsible for showing the log in form.
The method is a helper for setting error messages when a user is
redirected back after a failed log in attempt. It is not for the default
`$Form` output in templates. Methods external to SAMLController should
fetch the form from the Authenticator, or directly via Injector if they
need it.

SAMLHelper has had a refactor and comments added because it was rather
confusing.

Many of the tests added are seeminly unnecessary (are of little value
e.g. ErrorPageStaticPublishTest), however much of the output relies on
arbitrary conformancy to a specification documented elsewhere (php-saml
library). While we are not testing this library, we are testing our
creation of config for it will be compatible (see SAMLConfiguration).

SAMLMiddleware had a slight refactor in order to make it a little easier
to test via a subclass rather than mocking - as mocking can't work with
non-public or static methods.