API Remove deprecated code from LoginHandler
SAMLLoginForm::__construct
--------------------------

$authenticationClass parameter removed
======================================

This appears to be of convention, rather than because of interface
compliance. The variable is of no use, and a major version has already
elapsed since it was marked as deprecated - it is time for it to go.

AuthenticationMethod form field
===============================

The only use of $authenticatorClass in other forms (not this one)
appears to be for setting the 'AuthenticationMethod' hidden form field.
Although present in Silverstripe Framework 4, it is a forgotten piece
from SS2 & 3, where the authenticator to load was determined by this
field (OpenId, MemberAuthenticator, etc.)

SS4 completely rearchitected the authentication interfaces, and now an
Authenticator sets up the AuthenticationHandler to post to itself on the
log-in page's link. I.e. `Security/login/LoginForm` became
`Security/login/METHOD/LoginForm`.

Because of this both the $authenticatorClass constructor parameter and
AuthenticatorMethod field that used it have been removed.

$authenticator_class property
=============================

The procted $authenticatorClass property however remains, simply for
correctness where it is used by the LoginForm superclass's
getAuthenticator method. This property was renamed from
authenticator_class as has not been used since CMS 4 implemented PSR-2
(before it's first release).