FIX log in appearing to fail due to bad redirect
The default for the BackURL if it was not provided was `null`. However
the onelogin php-saml library's Auth method (that does the redirecting
to the IdP with the SAML request) would use $_SERVER variable to get the
current request URI if the provided RelayState was empty.
(RelayState is the SAML equivalent of Silverstripe's BackURL - for which
$backURL was used)

As we've submitted a form to perform the login action (SAMLLoginForm)
the current URL is the form submission handler for authentication -
which requires POST method (and data) and is not intended for
"browsing". The result of this flow would be an error reported from
Silverstripe CMS after a successful log in at the IdP. It is not obvious
in any way that the authentication was successful.

By setting the BackURL to $this->Link() if it is empty before we pass to
the library for authentication redirection we can have more control over
where the user will end up after the process completes.

I.e. In the very least, go back to the page we submitted the log in form
from.

See OneLogin\Saml2\Auth::login(), line 556 in v4.2.0
and OneLogin\Saml2\Utils::getSelfRoutedURLNoQuery()