Add convenience config for setting display name
When the log in form is presented to users, the default is displayed as
"SAML", which is rather unintuitive for people who don't know what SAML
is, or which provider it is using for the IdP. It is translatable, which
is currently the only way to change this to state "Staff log in"
(presuming a company's Active Directory as the IdP in this example).

This introduces a minor barrier in that one must create language files
and correctly format them for the SAML authenticator class in order to
change the name (and _not_ translate it). As this is distinct from
translation and presuming a single language site, a configurable
fallback for the translation default makes sense - allowing an one to
easily change the title/name without having to delve into language
configurations.

This option in place means that he existing language file entry for the
authenticator name must be removed in order for it to work.