FIX assign all groups from trusted source
Programmatic assigning of groups as delivered signed and potentially
encrypted from a SAML IdP was being blocked by Silverstripe core in
certain critical conditions.

The first step in the user mapping feature is to remove all groups to
short-cut revoking access where no longer needed. This then results in
an ADMIN no longer having that permission. Adding a group to a member
has a silent skip when the group has ADMIN permissions, and both the
target and the assigning user don't.

This is a problem when we've just stripped ADMIN privilges from a user
that would otherwise be one. It also prevents the first users in a
system from having rightly assigned ADMIN access without having to use a
break-glass account as a workaround to assign it.

Since we should be able to trust our external source, and programmatic
execution does not have a corresponding member to check against, we can
instead work-around the privilige escalation check/block by simply
assinging the member to a group, instead of a group to the member.